Add `validate_sql_json` endpoint for checking that a given sql query is valid for the chosen database (#7422)
* Add "validation_only" queries to the backend

This is the first step toward supporting live as-you-type
SQL validation against databases that support doing it (for
starters just presto).

I'm trying to stick as much to the existing query machinery
as possible to avoid introducing too much new state for handling
this, so I plumbed it through as a db spec parameter and an
option you can set on an individual query.

As of this commit:

- [X] Plumb validation db_spec options
- [X] Add validation_only option and transform to the sql_json view
- [ ] Tests for the above on the backend
- [ ] Debounced requestValidationQuery in onSqlChange
- [ ] Check incoming results -- full query vs validation
- [ ] Tests for the above on the frontend
- [ ] Highlight error lines
- [ ] Print error message in alert-danger box above toolbar controls

* Review comments, phase 1.

Factors out the validation process to a separate class which
does validation inline, rather than passing it through the
existing query flow implicitly.

This is meant to address Dave's feedback requesting that
the validation flow not be explicitly tied to a query transform
since that's uniquely a presto-ism.

Next up in this stack: unit tests.

* Split validation off to a separate endpoint

Just to demo it and see if it's cleaner this wak.

* Add unit tests covering validation

The new validator files are >90% covered, and the coverage
on the core view is back up to about where it was before this
changeset.

* Fix flake8 lints

`tox -e flake8` is clean now

* Hacky test fix: bind mock validator for pg/mysql

* Fix pylint lints

Also fix the flake8s that showed up from fixing pylint.

* Use 400 Bad Request for validation user-error

Consider it user-error and respond with 400 instead of 500
if somebody requests validation with template parameters,
for a backend with no configured validator, or for an
invalid database ID.

These are in contrast to the actual server error reply for
exceptions thrown by the underlying validator implementation,
which still (IMO correctly) returns 500.

* Move validator config to a feature flag

* Add docs on configuring validators

* Fix linter complaints

* Extend DatabaseError catch block and document it

The pyhive presto client yields EXPLAIN (TYPE VALIDATE)
responses as though they were normal queries -- in other
words it doesn't know that errors for that kind of request
aren't exceptional. To work around that, we have to trap
the exceptions it returns and translate the ones that
express problems with the SQL query.

* fix postgresql test environment

* Hit pylint with a hammer

It's convinced these are the wrong type. They aren't.

* More lint fixes