Added in support for ip_string and ability to emit
I don't think we will ever emit from AAG but I did manage to get it to
work so I'll push those changes.

If a filter comes in for ip_string it will filter AAG.  The filter has
to be for ip_string and must contain an array of values with at least
one value.

Updated the style of AAG so that it will fit nicely at the top of a
dashboard with tabs for other GWWK pieces below it.